refactor: Controller 제거를 위한 패키지 구조 변경
