+++ .gitignore
@@ -44,3 +44,8 @@ docker/db/mysql/init/*
 /src/test/java/chess/dao/MemberDaoTest.java
 /src/main/java/chess/Role.java
 /src/test/java/chess/dao/
+
+
+/src/main/java/chess/MyApplication.java
+/src/main/java/chess/MyCommand.java
+/src/main/java/chess/Cf.java
