+++ src/main/java/chess/domain/gamecommand/Command.java
@@ -1,8 +1,8 @@
-package chess.controller.command;
+package chess.domain.gamecommand;
 
 import chess.view.OutputView;
 
 public interface Command {
-    
+
     void execute(String rawInputCommand, final OutputView outputView);
 }
