+++ src/main/java/chess/domain/gamecommand/Move.java
@@ -1,11 +1,11 @@
-package chess.controller.command;
+package chess.domain.gamecommand;
 
 import chess.domain.ChessGame;
 import chess.domain.board.Positions;
 import chess.view.OutputView;
 
 public class Move implements Command {
-    
+
     private final ChessGame chessGame;
 
     public Move(final ChessGame chessGame) {
