+++ src/main/java/chess/domain/gamecommand/Start.java
@@ -1,10 +1,10 @@
-package chess.controller.command;
+package chess.domain.gamecommand;
 
 import chess.domain.ChessGame;
 import chess.view.OutputView;
 
 public class Start implements Command {
-    
+
     private final ChessGame chessGame;
 
     public Start(final ChessGame chessGame) {
