+++ src/main/java/chess/domain/gamecommand/Status.java
@@ -1,10 +1,10 @@
-package chess.controller.command;
+package chess.domain.gamecommand;
 
 import chess.domain.ChessGame;
 import chess.view.OutputView;
 
 public class Status implements Command {
-    
+
     private final ChessGame chessGame;
 
     public Status(final ChessGame chessGame) {
